fixed audit date bug - autoboxing error, mixing int and long in multiplication statement results in negative number when all multiplicands/factors.

+++ src/main/java/org/dataone/service/cn/replication/auditor/v1/controller/CoordinatingNodeReplicationAuditor.java
@@ -46,7 +46,7 @@ public class CoordinatingNodeReplicationAuditor extends AbstractReplicationAudit
     private static final int taskPoolSize = 10;
     private static final int maxPages = 50;
 
-    private static final int auditPeriodDays = Settings.getConfiguration().getInt(
+    private static final long auditPeriodDays = Settings.getConfiguration().getLong(
             "Replication.audit.cn.period.days", 90);
 
     private static final long auditPeriod = 1000 * 60 * 60 * 24 * auditPeriodDays;
